+++ src/lib/main.dart
@@ -1,5 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';
+
+import 'sign_in_page.dart';
 
 void main() => runApp(const MyApp());
 
@@ -39,25 +42,27 @@ class HomePage extends StatefulWidget {
 
 class _HomePageState extends State<HomePage> {
   int _selectedIndex = 1;
-  static const List<Widget> _widgetOptions = <Widget>[
-    Text(
-      'Index 0: Account',
-      style: TextStyle(
-        fontFamily: 'CascadiaMono',
-        fontSize: 50,
-      ),
-    ),
-    AddFundraiserWidget(),
-    Text(
-      'Index 2: Logout',
-    ),
-  ];
 
   void _onItemTapped(int index) {
     setState(() {
       _selectedIndex = index;
     });
+
+      if (index == 2) {
+      Navigator.push(
+        context,
+        MaterialPageRoute(builder: (context) => SignInPage()),
+      );
+    }
   }
+
+  static List<Widget> _widgetOptions = <Widget>[
+    Text(
+      'Index 0: Account',),
+    AddFundraiserWidget(),
+    Text('',),
+  ];
+
 
   @override
   Widget build(BuildContext context) {
@@ -80,188 +85,6 @@ class _HomePageState extends State<HomePage> {
         selectedItemColor: theme.colorScheme.primary,
         onTap: _onItemTapped,
       ),
-    );
-  }
-}
-
-class SignInPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      margin: EdgeInsets.all(16.0),
-      elevation: 4.0,
-      child: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            _buildTitle(context),
-            SizedBox(height: 16.0),
-            _buildTextField(hintText: 'Email', icon: Icons.email),
-            SizedBox(height: 16.0),
-            _buildTextField(
-                hintText: 'Password', icon: Icons.lock, obscureText: true),
-            SizedBox(height: 16.0),
-            _buildButtons(context),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildTitle(BuildContext context) {
-    final theme = Theme.of(context);
-
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        Text(
-          'Sign In',
-          style: TextStyle(
-            fontSize: 24.0,
-            fontWeight: FontWeight.bold,
-          ),
-          textAlign: TextAlign.center,
-        ),
-        TextButton(
-          onPressed: () {
-            Navigator.push(
-              context,
-              MaterialPageRoute(builder: (context) => SignUpPage()),
-            );
-          },
-          child: Text(
-            'Sign Up',
-            style: TextStyle(
-              fontSize: 16.0,
-              color: theme.colorScheme.primary,
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildTextField(
-      {required String hintText,
-      required IconData icon,
-      bool obscureText = false}) {
-    return TextField(
-      obscureText: obscureText,
-      decoration: InputDecoration(
-        hintText: hintText,
-        prefixIcon: Icon(icon),
-        border: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(8.0),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildButtons(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-      children: [
-        ElevatedButton(
-          onPressed: () {
-            Navigator.push(
-            context,
-            MaterialPageRoute(builder: (context) => HomePage()),
-          );
-          },
-          child: Text('Sign In'),
-        ),
-      ],
-    );
-  }
-}
-
-class SignUpPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      margin: EdgeInsets.all(16.0),
-      elevation: 4.0,
-      child: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            _buildTitle(context),
-            SizedBox(height: 16.0),
-            _buildTextField(hintText: 'Username', icon: Icons.person),
-            SizedBox(height: 16.0),
-            _buildTextField(hintText: 'Email', icon: Icons.email),
-            SizedBox(height: 16.0),
-            _buildTextField(
-                hintText: 'Password', icon: Icons.lock, obscureText: true),
-            SizedBox(height: 16.0),
-            _buildButtons(context),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildTitle(BuildContext context) {
-    final theme = Theme.of(context);
-
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        Text(
-          'Sign Up',
-          style: TextStyle(fontSize: 24.0, fontWeight: FontWeight.bold),
-        ),
-        TextButton(
-          onPressed: () {
-            Navigator.push(
-              context,
-              MaterialPageRoute(builder: (context) => SignInPage()),
-            );
-          },
-          child: Text(
-            'Login',
-            style: TextStyle(
-              fontSize: 16.0,
-              color: theme.colorScheme.primary,
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildTextField(
-      {required String hintText,
-      required IconData icon,
-      bool obscureText = false}) {
-    return TextField(
-      obscureText: obscureText,
-      decoration: InputDecoration(
-        hintText: hintText,
-        prefixIcon: Icon(icon),
-        border: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(8.0),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildButtons(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-      children: [
-        ElevatedButton(
-          onPressed: () {
-            Navigator.push(
-              context,
-              MaterialPageRoute(builder: (context) => SignInPage()),
-            );
-          },
-          child: Text('Sign Up'),
-        ),
-      ],
     );
   }
 }
@@ -289,7 +112,7 @@ class AddFundraiserWidget extends StatelessWidget {
           child: TextFormField(
             decoration: InputDecoration(
               border: OutlineInputBorder(),
-              hintText: 'Title',
+              labelText: 'Title',
               fillColor: Color.fromARGB(240, 240, 240, 240),
               filled: true,
             ),
